프로젝트 열었을때 바로 main 화면으로 가게 하기 + 화면 너비 디자인 전체적으로 수정
