+++ src/main/java/com/project/Booktion/controller/auctionBook/AuctionBookListController.java
@@ -1,6 +1,8 @@
 package com.project.Booktion.controller.auctionBook;
 
 import com.project.Booktion.model.AuctionBook;
+import com.project.Booktion.model.AuctionBookOrder;
+import com.project.Booktion.repository.AuctionBookOrderRepository;
 import com.project.Booktion.repository.AuctionBookRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
@@ -17,6 +19,7 @@ import java.util.List;
 public class AuctionBookListController {
 
     private final AuctionBookRepository auctionBookR;
+    private final AuctionBookOrderRepository auctionOrderR;
 
     @GetMapping
     public String bookList(Model model){ // 전체 경매 책 불러오기
@@ -29,15 +32,18 @@ public class AuctionBookListController {
     @GetMapping("/selling") // userId 읽고 user가 경매중인 책
     public String sellingList(HttpSession session, Model model){
         String userId = (String) session.getAttribute("userId");
-        List<AuctionBook> sellingBooks = auctionBookR.findBySellerId(userId);
+
+        List<AuctionBook> sellingBooks = auctionBookR.findByUserSelling(userId);
+        // 반환된 책들중 id를 검색해서 oder가 잇는지 보고 order가 없는 것만 반환
+
         model.addAttribute("auctionBooks", sellingBooks);
-        return "myPage/selling";
+        return "myPage/auction/selling";
     }
     @GetMapping("/sold") //  userId 읽고 user가 경매 완료 책
     public String SoldList(HttpSession session, Model model){
         String userId = (String) session.getAttribute("userId");
-        List<AuctionBook> soldBooks = auctionBookR.findBySellerId(userId);
-        model.addAttribute("sellingBooks", soldBooks);
-        return "myPage/sold";
+        List<AuctionBookOrder> auctionBookOrders = auctionOrderR.findByUserIdSold(userId);
+        model.addAttribute("auctionBookOrders", auctionBookOrders);
+        return "myPage/auction/sold";
     }
 }
