+++ src/main/java/com/project/Booktion/controller/user/LoginFormController.java
@@ -30,7 +30,7 @@ public class LoginFormController {
         if (userRepository.findByUserIdAndPassword(userId, password) != null) {
             session.setAttribute("userId",userId);
             model.addAttribute("success","로그인에 성공했습니다");
-            return "redirect:/main"; // 로그인 성공 시 리다이렉트할 URL
+            return "redirect:/"; // 로그인 성공 시 리다이렉트할 URL
         } else {
             model.addAttribute("error", "로그인에 실패하였습니다\n다시 시도해 주세요");
             return "user/signIn"; // 로그인 폼으로 돌아가기
