+++ src/main/java/com/project/Booktion/repository/AuctionBookOrderRepository.java
@@ -11,8 +11,7 @@ import java.util.List;
 public interface AuctionBookOrderRepository extends JpaRepository<AuctionBookOrder, Long> {
 
     //내가 팔고있는 auctionBook을 검색하려 AuctionBookOrder로 반환
-    List<AuctionBookOrder> findByOrderUserUserId(@Param("userId") String userId);
-
-
-
+    List<AuctionBookOrder> findByOrderUserUserId(String userId);
+    @Query("SELECT abo FROM AuctionBookOrder abo JOIN abo.auctionBook.book b JOIN b.user u WHERE u.userId = :userId")
+    List<AuctionBookOrder> findByUserIdSold(@Param("userId") String userId);
 }
