+++ src/main/java/com/project/Booktion/repository/AuctionBookRepository.java
@@ -1,6 +1,7 @@
 package com.project.Booktion.repository;
 
 import com.project.Booktion.model.AuctionBook;
+import com.project.Booktion.model.AuctionBookOrder;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -11,6 +12,6 @@ import java.util.List;
 public interface AuctionBookRepository extends JpaRepository<AuctionBook, Long>{
 
     // userId로 acutionBook 검색
-    @Query(value = "SELECT ab FROM AuctionBook ab JOIN ab.book b WHERE b.user.userId = :sellerId", nativeQuery = true)
-    List<AuctionBook> findBySellerId(@Param("sellerId") String sellerId);
+    @Query("SELECT ab FROM AuctionBook ab WHERE ab.book.user.userId = :userId AND ab.status = 0")
+    List<AuctionBook> findByUserSelling(@Param("userId") String userId);
 }
