+++ src/main/resources/application.yml
@@ -23,6 +23,7 @@ spring:
    interceptor:
     include: /**
 
+
 logging:
   level:
     org.hibernate.SQL: debug
