+++ src/main/resources/templates/auction/book.html
@@ -9,7 +9,7 @@
     <link href="https://fonts.googleapis.com/css?family=Nanum+Gothic" rel="stylesheet">
     <style>
         .container {
-            max-width: 800px;
+            max-width: 1000px;
         }
 
         .card {
