+++ src/main/resources/templates/common/sidebar.html
@@ -7,9 +7,9 @@
         <h4 class="page-title">마이페이지</h4>
         <p class="section-title">주문 목록</p>
         <ul class="book-links">
-            <li><a class="side" href="#">일반도서 주문</a></li>
+            <li><a class="side" th:href="@{/myPage/order/basicBook}">일반도서 주문</a></li>
             <li><a class="side" th:href="@{/myPage/order/used}">중고도서 주문</a></li>
-            <li><a class="side" th:href="@{/auction/order/list}">경매도서 주문</a></li>
+            <li><a class="side" th:href="@{/myPage/order/auction}">경매도서 주문</a></li>
         </ul>
 
         <p class="section-title">판매하는 도서</p>
@@ -23,11 +23,11 @@
         <div class="sub-section">
             <p>경매도서</p>
             <ul class="book-links">
-                <li><a class="side" href="#">판매 중인 도서</a></li>
-                <li><a class="side" href="#">판매 완료 도서</a></li>
+                <li><a class="side" th:href="@{/auction/books/selling}">판매 중인 도서</a></li>
+                <li><a class="side" th:href="@{/auction/books/sold}">판매 완료 도서</a></li>
             </ul>
         </div>
 
-        <p><a href="#" class="side" style="text-decoration: none;">리뷰 작성하기</a></p>
+        <p><a class="side" th:href="@{/review}">작성한 리뷰</a></p>
     </div>
 </div>
