+++ src/main/resources/templates/main.html
@@ -22,7 +22,6 @@
 <a href="/used/main">중고도서</a><br>
 
 <a href="/book/list">책 목록</a>
-<a href="/search/auction">도서 검색 api -  경매</a>
 <a href="/search/used">도서 검색 api -  중고</a>
 <br>
 <a href="/review">리뷰</a><br>
