+++ src/main/resources/templates/myPage/auction/auctionOrderList.html
@@ -4,14 +4,14 @@
       layout:decorate="~{layout/layout}">
 <head>
     <meta charset="UTF-8">
-    <title>자세히보기</title>
+    <title>주문한 경매 도서</title>
     <link href="https://fonts.googleapis.com/css?family=Nanum+Gothic" rel="stylesheet">
     <style>
         .container {
             display: flex;
             justify-content: center;
             align-items: flex-start;
-            width: 900px;
+            width: 1000px;
             margin: 0 auto;
         }
         body {
@@ -21,7 +21,7 @@
             padding: 0;
         }
         .myPageContent {
-            width: 800px;
+            width: 860px;
             padding: 20px;
         }
         .auction-book-orders {
