+++ src/main/resources/templates/myPage/auction/detailOrder.html
@@ -11,7 +11,7 @@
             display: flex;
             justify-content: center;
             align-items: flex-start;
-            width: 900px;
+            width: 1000px;
             margin: 0 auto;
         }
         body {
