+++ src/main/resources/templates/myPage/review/reviewList.html
@@ -1,57 +1,126 @@
 <!DOCTYPE html>
-<html lang="ko" xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout" layout:decorate="~{layout/layout}">
+<html lang="ko" xmlns:th="http://www.thymeleaf.org"
+      xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+      layout:decorate="~{layout/layout}">
 <head>
     <meta charset="UTF-8">
     <title>리뷰</title>
+    <link href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css?family=Nanum+Gothic" rel="stylesheet">
     <style>
+        .container {
+            display: flex;
+            justify-content: center;
+            align-items: flex-start;
+            width: 1000px;
+            margin: 0 auto;
+            padding: 30px;
+        }
+        body {
+            font-family: 'Nanum Gothic', sans-serif;
+            background-color: #ffffff;
+            margin: 0;
+            padding: 0;
+        }
+        .h3 {
+            font-size: 24px;
+            margin-bottom: 10px;
+        }
+        .myPageContent {
+            width: 860px;
+            padding: 20px;
+        }
         table {
-              width: 100%;
-              border-collapse: collapse;
-              margin-bottom: 20px;
-            }
+            width: 100%;
+            border-collapse: collapse;
+            margin-top: 10px;
+        }
         th, td {
-            border: 1px solid black;
+            border: 1px solid #b2d8e5;
             padding: 8px;
+            text-align: left;
         }
         th {
-            background-color: #f2f2f2;
+            background-color: #b2d8e5;
+            font-weight: bold;
+        }
+        a {
+            color: #2892c6;
+            text-decoration: none;
+        }
+        form {
+            margin: 0;
+        }
+        button {
+            background-color: transparent;
+            border: none;
+            color: #2892c6;
+            cursor: pointer;
+        }
+        .write-review-button {
+            display: inline-block;
+            padding: 8px 20px;
+            background-color: #b2d8e5;
+            border: 4px solid #b2d8e5;
+            border-radius: 8px;
+            color: #ffffff;
+            text-decoration: none;
+        }
+        .write-review-button:hover {
+            background-color: #2892c6;
+            border-color: #2892c6;
+        }
+        .write-review-button-container {
+
+            text-align: right;
         }
     </style>
 </head>
 <body>
+<br><br><br>
 <div layout:fragment="content">
-    <h1>리뷰리스트</h1>
-    <table>
-        <thead>
-        <tr>
-            <th> </th>
-            <th>글 번호</th>
-            <th>책 이름</th>
-            <th>제목</th>
-            <th>내용</th>
-            <th>날짜</th>
-            <th> </th>
-        </tr>
-        </thead>
-        <tbody>
-        <tr th:each="review : ${reviews}">
-            <td>
-                <p>
-                    <img th:src="${review.book.imageUrl}" alt="Book Image" />
-                </p>
-            </td>
-            <td th:text="${review.reviewId}"></td>
-            <td th:text="${review.book.title}"></td>
-            <td th:text="${review.title}"></td>
-            <td th:text="${review.contents}"></td>
-            <td th:text="${review.createDate}"></td>
-            <td>
-                <a th:href="@{/review/editReview/{reviewId}(reviewId=${review.reviewId})}">수정</a>
-            </td>
-        </tr>
-        </tbody>
-    </table>
-    <a th:href="@{/review/createReview}">리뷰 작성하기</a>
+    <br>
+    <div class="container">
+        <div class="sidebar-wrap">
+            <th:block th:replace="common/sidebar :: SidebarFragment"></th:block>
+        </div>
+        <div class="myPageContent">
+            <h3>리뷰리스트</h3>
+            <table>
+                <thead>
+                <tr>
+                    <th> </th>
+                    <th>글 번호</th>
+                    <th>책 이름</th>
+                    <th>제목</th>
+                    <th>내용</th>
+                    <th>날짜</th>
+                    <th> </th>
+                </tr>
+                </thead>
+                <tbody>
+                <tr th:each="review : ${reviews}">
+                    <td>
+                        <p>
+                            <img th:src="${review.book.imageUrl}" alt="Book Image" />
+                        </p>
+                    </td>
+                    <td th:text="${review.reviewId}"></td>
+                    <td th:text="${review.book.title}"></td>
+                    <td th:text="${review.title}"></td>
+                    <td th:text="${review.contents}"></td>
+                    <td th:text="${review.createDate}"></td>
+                    <td>
+                        <a th:href="@{/review/editReview/{reviewId}(reviewId=${review.reviewId})}">수정</a>
+                    </td>
+                </tr>
+                </tbody>
+            </table>
+            <div class="write-review-button-container">
+                <a class="write-review-button" th:href="@{/review/createReview}">리뷰 작성하기</a>
+            </div>
+        </div>
+    </div>
 </div>
 </body>
 </html>
