+++ src/main/resources/templates/myPage/used/detailOrder.html
@@ -11,7 +11,7 @@
             display: flex;
             justify-content: center;
             align-items: flex-start;
-            width: 900px;
+            width: 1000px;
             margin: 0 auto;
         }
         body {
@@ -21,7 +21,7 @@
             padding: 0;
         }
         .myPageContent {
-            width: 700px;
+            width: 860px;
             padding: 20px;
         }
